@@ -1,50 +1,37 @@
 {
-  "name": "ngel-product-page",
-  "version": "9.0.0",
+  "name": "meghsikshak-product-page",
   "private": true,
-  "dependencies": {
-    "@fortawesome/fontawesome-free": "^6.7.0",
-    "@testing-library/jest-dom": "^5.16.5",
-    "@testing-library/react": "^13.4.0",
-    "@testing-library/user-event": "^13.5.0",
-    "daisyui": "^2.51.1",
-    "lucide-react": "^0.460.0",
-    "ngel-product-page": "file:",
-    "react": "^18.2.0",
-    "react-countup": "^6.5.3",
-    "react-dom": "^18.2.0",
-    "react-icons": "^5.3.0",
-    "react-scripts": "5.0.1",
-    "react-slick": "^0.30.2",
-    "slick-carousel": "^1.8.1",
-    "swiper": "^11.1.15",
-    "web-vitals": "^2.1.4"
-  },
+  "version": "10.0.0",
+  "type": "module",
   "scripts": {
-    "start": "set PORT=5000 && react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject"
+    "dev": "vite",
+    "build": "vite build",
+    "lint": "eslint .",
+    "preview": "vite preview"
   },
-  "eslintConfig": {
-    "extends": [
-      "react-app",
-      "react-app/jest"
-    ]
-  },
-  "browserslist": {
-    "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ]
+  "dependencies": {
+    "@fortawesome/fontawesome-free": "^6.7.1",
+    "lucide-react": "^0.460.0",
+    "meghsikshak-product-page": "file:",
+    "meghsikshak-product-page-vite": "file:",
+    "react": "^18.3.1",
+    "react-countup": "^6.5.3",
+    "react-dom": "^18.3.1"
   },
   "devDependencies": {
-    "tailwindcss": "^3.2.7"
+    "@eslint/js": "^9.13.0",
+    "@types/react": "^18.3.12",
+    "@types/react-dom": "^18.3.1",
+    "@vitejs/plugin-react": "^4.3.3",
+    "autoprefixer": "^10.4.20",
+    "daisyui": "^4.12.14",
+    "eslint": "^9.13.0",
+    "eslint-plugin-react": "^7.37.2",
+    "eslint-plugin-react-hooks": "^5.0.0",
+    "eslint-plugin-react-refresh": "^0.4.14",
+    "globals": "^15.11.0",
+    "postcss": "^8.4.49",
+    "tailwindcss": "^3.4.15",
+    "vite": "^5.4.10"
   }
 }
